requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
boto3==1.34.4
botocore==1.34.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.8
django-allauth==0.59.0
django-auto-logout==0.5.1
django-bootstrap-form==3.4
django-bootstrap4==23.2
django-crispy-forms==2.1
django-storages==1.14.2
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.9.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==1.25.4
whitenoise==6.6.0
